+++ src/reuse/init.py
@@ -4,6 +4,7 @@
 
 """Functions for REUSE-ifying a project."""
 
+import re
 import sys
 from argparse import ArgumentParser, Namespace
 from gettext import gettext as _
@@ -13,7 +14,11 @@ from typing import IO, List
 from urllib.error import URLError
 
 from ._licenses import ALL_NON_DEPRECATED_MAP
-from ._util import PathType, print_incorrect_spdx_identifier
+from ._util import (
+    _LICENSEREF_PATTERN,
+    PathType,
+    print_incorrect_spdx_identifier,
+)
 from .download import _path_to_license_file, put_license_in_file
 from .project import Project
 from .vcs import find_root
@@ -44,7 +49,9 @@ def prompt_licenses(out: IO[str] = sys.stdout) -> List[str]:
         out.write("\n")
         if not result:
             return licenses
-        if result not in ALL_NON_DEPRECATED_MAP:
+        if result not in ALL_NON_DEPRECATED_MAP and not re.match(
+            _LICENSEREF_PATTERN, result
+        ):
             print_incorrect_spdx_identifier(result, out=out)
             out.write("\n\n")
         else:
@@ -116,8 +123,9 @@ def run(
 
     for lic in licenses:
         destination = _path_to_license_file(lic, root=root)
+
         try:
-            out.write(_("Downloading {}").format(lic))
+            out.write(_("Retrieving {}").format(lic))
             out.write("\n")
             put_license_in_file(lic, destination=destination)
         # TODO: exceptions
@@ -126,6 +134,14 @@ def run(
             out.write("\n")
         except URLError:
             out.write(_("Could not download {}").format(lic))
+            out.write("\n")
+        except FileNotFoundError as err:
+            out.write(
+                _(
+                    "Error: Could not copy {path}, "
+                    "please add {lic}.txt manually in the LICENCES/ directory."
+                ).format(path=err.filename, lic=lic)
+            )
             out.write("\n")
 
     out.write("\n")